Rename scaleCoordinate.js to axesUtils.js and add scalePartialCoordinate
